pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zundamotion"
dynamic = ["version"]
description = "Generate videos from YAML scripts using VOICEVOX and FFmpeg."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "requests==2.32.4",
  "PyYAML==6.0.2",
  "pydub==0.25.1",
  "tqdm==4.67.1",
  "pysubs2==1.8.0",
  "Pillow==11.3.0",
  "httpx==0.28.1",
  "tenacity==9.1.2",
  "rembg==2.0.67",
  "onnxruntime==1.22.1",
]

[project.optional-dependencies]
dev = ["pytest>=7"]

[project.scripts]
zundamotion = "zundamotion.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { attr = "zundamotion.__version__" }

[tool.setuptools.packages.find]
include = ["zundamotion*"]

[tool.setuptools.package-data]
zundamotion = ["templates/*.yaml", "plugins/builtin/*/plugin.yaml"]

